Fixed MAGE-6567: Payment Bridge: After clicking button "Submit", Payment Bridge validation form doesn't displayed

+++ app/code/core/Mage/Core/Controller/Varien/Front.php
@@ -297,7 +297,7 @@ class Mage_Core_Controller_Varien_Front extends Varien_Object
             return;
         }
 
-        $redirectCode = Mage::getStoreConfig('web/url/redirect_to_base');
+        $redirectCode = (int)Mage::getStoreConfig('web/url/redirect_to_base');
         if (!$redirectCode) {
             return;
         } elseif ($redirectCode != 301) {
